--- a/packages/bbui/src/DetailSummary/DetailSummary.svelte
+++ b/packages/bbui/src/DetailSummary/DetailSummary.svelte
@@ -2,12 +2,10 @@
   import Icon from "../Icon/Icon.svelte"
   import { createEventDispatcher } from "svelte"
 
+  export let name
+  export let show = false
+
   const dispatch = createEventDispatcher()
-
-  export let thin = false
-  export let name,
-    show = false
-
   const onHeaderClick = () => {
     show = !show
     if (show) {
@@ -16,12 +14,10 @@
   }
 </script>
 
-<div class="property-group-container" class:thin>
+<div class="property-group-container">
   <div class="property-group-name" on:click={onHeaderClick}>
-    <div class:thin class="name">{name}</div>
-    <div class="icon">
-      <Icon size="S" name={show ? "Remove" : "Add"} />
-    </div>
+    <div class="name">{name}</div>
+    <Icon size="S" name={show ? "Remove" : "Add"} />
   </div>
   <div class="property-panel" class:show>
     <slot />
@@ -32,10 +28,9 @@
   .property-group-container {
     display: flex;
     flex-direction: column;
-    height: auto;
-    justify-content: center;
-    border-radius: var(--border-radius-m);
-    font-family: var(--font-sans);
+    justify-content: flex-start;
+    align-items: stretch;
+    border-bottom: var(--border-light);
   }
 
   .property-group-name {
@@ -45,42 +40,38 @@
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
+    padding: var(--spacing-m) var(--spacing-xl);
+    color: var(--spectrum-global-color-gray-600);
+    transition: color 130ms ease-in-out;
+  }
+  .property-group-name:hover {
+    color: var(--spectrum-global-color-gray-900);
   }
 
   .name {
     text-align: left;
-    font-size: 14px;
+    font-size: var(--font-size-s);
     font-weight: 600;
     letter-spacing: 0.14px;
-    color: var(--ink);
     flex: 1 1 auto;
     overflow: hidden;
     text-overflow: ellipsis;
-    text-transform: capitalize;
+    text-transform: uppercase;
     white-space: nowrap;
     user-select: none;
   }
-  .name.thin {
-    font-size: var(--spectrum-global-dimension-font-size-75);
-  }
-
-  .icon {
-    flex: 0 0 20px;
-    text-align: center;
-  }
 
   .property-panel {
-    /* height: 0px;
-    overflow: hidden; */
     display: none;
+    padding: var(--spacing-s) var(--spacing-xl) var(--spacing-xl)
+      var(--spacing-xl);
   }
 
   .show {
-    /* overflow: auto;
-    height: auto; */
     display: flex;
     flex-direction: column;
-    flex: 1;
-    margin-top: var(--spacing-m);
+    justify-content: flex-start;
+    align-items: stretch;
+    gap: var(--spacing-m);
   }
 </style>