… packages/builder/src/components/design/PropertiesPanel/CategoryTab.svelte
… packages/builder/src/components/design/PropertiesPanel/CustomStylesSection.svelte
… packages/builder/src/components/design/PropertiesPanel/DesignSection.svelte
… packages/builder/src/components/design/PropertiesPanel/DesignView.svelte
… packages/builder/src/components/design/PropertiesPanel/PropertiesPanel.svelte
… packages/builder/src/components/design/PropertiesPanel/PropertyControls/PropertyGroup.svelte
… packages/builder/src/components/design/PropertiesPanel/PropertyControls/ResetFieldsButton.svelte
… packages/builder/src/components/design/PropertiesPanel/ScreenSettingsSection.svelte
… packages/builder/src/components/design/PropertiesPanel/SettingsSection.svelte
… packages/builder/src/components/design/PropertiesPanel/StyleSection.svelte
… packages/builder/src/components/design/PropertiesPanel/componentStyles.js
… packages/builder/src/pages/builder/app/[application]/data/datasource/[selectedDatasource]/index.svelte
… packages/builder/src/pages/builder/app/[application]/data/datasource/bb_internal/index.svelte
… packages/client/src/components/ClientApp.svelte
… packages/client/src/components/Component.svelte
… packages/client/src/components/Provider.svelte
… packages/client/src/components/preview/SettingsBar.svelte
… packages/client/src/components/preview/SettingsButton.svelte
… packages/client/src/components/preview/SettingsColorPicker.svelte
… packages/client/src/components/preview/SettingsPicker.svelte
… packages/client/src/store/builder.js
… packages/client/src/store/context.js
… packages/client/src/utils/buttonActions.js
… packages/client/src/utils/componentProps.js
… packages/client/src/utils/styleable.js
… packages/server/src/api/controllers/layout.js
… packages/server/src/constants/layouts.js
… packages/server/src/constants/screens.js
… packages/server/yarn.lock
… packages/standard-components/manifest.json