## Description
This PR primarily contains the work for the unification of the settings and design panels in to one simplified, improved panel. A lot of styles and bloat have been removed, and now only the styles that are relevant to particular components are displayed. All components have at least one style setting - margin, and all still have custom styles (which now opens in a drawer).

This PR also contains huge performance improvements for client apps, particularly when interacting with the builder preview.

There are also a bunch of miscellaeneous features and improvements.

### Design panel V2
![image](https://user-images.githubusercontent.com/9075550/123451897-d304f580-d5d5-11eb-9833-f73ec922b4d2.png)
![image](https://user-images.githubusercontent.com/9075550/123455995-3e50c680-d5da-11eb-9f44-88e48be9dd62.png)

The design panel has settings at the top, with styles underneath. Each style "section" is now visible segmented with thin borders, and style components have be layed out in more compact ways.

For example, here's the style section for background color with the new color picker inline with a select.
![image](https://user-images.githubusercontent.com/9075550/123452269-28410700-d5d6-11eb-80c0-df3f9ad09939.png)

Which style sections are rendered for each component is controlled by the manifest, which means there is a single source of truth for all component settings.
![image](https://user-images.githubusercontent.com/9075550/123452528-676f5800-d5d6-11eb-9275-cce02688b8f1.png)

The layout of each style section is controlled in the existing `componentStyles.js` file, using grid layout. For example, here's the background section specifying to use custom column widths to achieve the correct layout. All style sections are still programmatically generated from these static definitions.
![image](https://user-images.githubusercontent.com/9075550/123453086-02683200-d5d7-11eb-9655-67e937be0e3e.png)

## Performance improvements
After experience an extremely slow builder experience when using my new PDF component with multiple pages (i.e. 100+ components on a screen) and seeing Joe's slow app, I felt it was important to try and improve performance. I did some benchmarking with both Firefox and Chrome devtools, and identified the handlebars enrichment of component props to be the cause of the nearly all the load. I think this is because of the number of enrichments, and the size of the context that is passed in.

To fix this, I've aggressively memoized a lot of state in the client library to avoid enriching components at all costs. The second part was identifying the slowest part of enrichment itself, which was the deep cloning of context and the multiple recursive functions to remove `nullish` values, and then removing these. These 2 changes combined have led to an almost 10X speed up in builder interactions in larger apps.

### Performance improvement in Firefox
Large app, PDF with multiple pages.
| Test | Current duration (ms) | New duration (ms) | Improvement |
| - | - | - | - |
| Select component in preview | 4300  | 40 | 108X |
| Change top-level screen prop | 4100 | 300 | 14X |
| First client app preview paint on page load (dev / vite) | 5700 | 3600 | 1.6X |

### Performance improvement in Chrome
Large app, PDF with multiple pages.
| Test | Current duration (ms) | New duration (ms) | Improvement |
| - | - | - | - |
| Select component in preview | 1100 | 40 | 28X |
| Change top-level screen prop | 1240 | 110 | 11X |
| First client app preview paint on page load (dev / vite) | 3800 | 2800 | 1.4X |

### Misc features
- Add new spectrum-themed color picker component
![image](https://user-images.githubusercontent.com/9075550/123455799-0c3f6480-d5da-11eb-93c5-cb7969c603b1.png)
- Add settings bar for typography settings for headings, paragraphs, links and repeaters
![image](https://user-images.githubusercontent.com/9075550/123455886-211bf800-d5da-11eb-8982-8b96989b2446.png)
- Add pretty labels to button options
- Add z index to layouts so that absolute elements can be added at the layout level without overlapping the nav bar
- Add placeholders to image, background image, embeds and link
- Remove "external" checkbox from links and work it out automatically
- Allow "navigate to" button action to go to external links
- Add "wrap" setting to containers
- Add container settings to repeaters, to allow easy customisation of repeater children layout