@mike12345567 2021-06-25
This is amazing! The new panel is so lovely, makes life so much easier! The performance improvements are incredible as well, can't believe the 108X improvement in FF - as well as the fact the select could be brought down to 40ms!